Learnings/BOOK-6-GO_CRUD: Created model package and used leads object in dao, service & controller.
