+++ services/leads.go
@@ -3,6 +3,7 @@ package services
 import (
 	"log"
 
+	"github.com/go4digital/booknow-api/model"
 	"github.com/go4digital/booknow-api/dao"
 	"github.com/go4digital/booknow-api/database"
 )
@@ -14,17 +15,17 @@ var (
 type leadService struct{}
 
 type leadServiceInterface interface {
-	CreateLead(*dao.Lead) (int64, error)
-	UpdateLead(*dao.Lead) (int64, error)
-	GetAllLeads() (*[]dao.Lead, error)
-	GetLead(int64) (*dao.Lead, error)
+	CreateLead(*model.Lead) (int64, error)
+	UpdateLead(*model.Lead) (int64, error)
+	GetAllLeads() (*[]model.Lead, error)
+	GetLead(int64) (*model.Lead, error)
 	DeleteLead(int64) (int64, error)
 }
 
 var db = database.Connect()
 var leads = dao.NewLeads(db)
 
-func (leadService *leadService) CreateLead(lead *dao.Lead) (int64, error) {
+func (leadService *leadService) CreateLead(lead *model.Lead) (int64, error) {
 	leadId, err := leads.CreateLead(lead)
 	if err != nil {
 		log.Println(err)
@@ -33,7 +34,7 @@ func (leadService *leadService) CreateLead(lead *dao.Lead) (int64, error) {
 
 }
 
-func (leadService *leadService) UpdateLead(lead *dao.Lead) (int64, error) {
+func (leadService *leadService) UpdateLead(lead *model.Lead) (int64, error) {
 	rowsAffected, err := leads.UpdateLead(lead)
 	if err != nil {
 		log.Println(err)
@@ -41,7 +42,7 @@ func (leadService *leadService) UpdateLead(lead *dao.Lead) (int64, error) {
 	return rowsAffected, err
 }
 
-func (leadService *leadService) GetAllLeads() (*[]dao.Lead, error) {
+func (leadService *leadService) GetAllLeads() (*[]model.Lead, error) {
 	leads, err := leads.GetAllLeads()
 	if err != nil {
 		log.Println(err)
@@ -49,7 +50,7 @@ func (leadService *leadService) GetAllLeads() (*[]dao.Lead, error) {
 	return leads, err
 }
 
-func (leadService *leadService) GetLead(leadId int64) (*dao.Lead, error) {
+func (leadService *leadService) GetLead(leadId int64) (*model.Lead, error) {
 	lead, err := leads.GetLead(leadId)
 	if err != nil {
 		log.Println(err)
